+++ facade/src/main/scala/react/semanticui/elements/flag/package.scala
@@ -2,4 +2,13 @@ package react.semanticui.elements
 
 package flag {}
 
-package object flag
+package object flag {
+  type UnmountedFlag = japgolly.scalajs.react.component.Js.UnmountedMapped[
+    japgolly.scalajs.react.internal.Effect.Id,
+    Flag.FlagProps,
+    Null,
+    japgolly.scalajs.react.component.Js.RawMounted[Flag.FlagProps, Null],
+    Flag.FlagProps,
+    Null]
+
+}