+++ facade/src/main/scala/react/semanticui/elements/image/package.scala
@@ -13,4 +13,13 @@ package image {
 
 }
 
-package object image
+package object image {
+
+  type UnmountedImage = japgolly.scalajs.react.component.Js.UnmountedMapped[
+    japgolly.scalajs.react.internal.Effect.Id,
+    Image.ImageProps,
+    Null,
+    japgolly.scalajs.react.component.Js.RawMounted[Image.ImageProps, Null],
+    Image.ImageProps,
+    Null]
+}